# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agentshield",
    "version": "0.0.1"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/version": {
      "get": {
        "summary": "Version",
        "operationId": "version_version_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Version Version Get"
                }
              }
            }
          }
        }
      }
    },
    "/runs": {
      "get": {
        "summary": "List Runs",
        "operationId": "list_runs_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunList"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Run",
        "operationId": "create_run_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}": {
      "get": {
        "summary": "Read Run",
        "operationId": "read_run_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/report": {
      "get": {
        "summary": "Read Report",
        "operationId": "read_report_runs__run_id__report_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/traces": {
      "get": {
        "summary": "Read Traces",
        "operationId": "read_traces_runs__run_id__traces_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunTraces"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{run_id}/diff/{other_id}": {
      "get": {
        "summary": "Diff Runs",
        "operationId": "diff_runs_runs__run_id__diff__other_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "other_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Other Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentTrace": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/TraceEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "final_text": {
            "type": "string",
            "title": "Final Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "events",
          "final_text"
        ],
        "title": "AgentTrace",
        "description": "The ordered record of a finished or stopped run."
      },
      "DiffBody": {
        "properties": {
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioDiff"
            },
            "type": "array",
            "title": "Scenarios"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scenarios"
        ],
        "title": "DiffBody",
        "description": "Scenario ids that changed outcome. `left` is the first run in the URL."
      },
      "FailingRule": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "severity",
          "reason"
        ],
        "title": "FailingRule",
        "description": "One rule that failed for a scenario."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ReportBody": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "scenario_pack_hash": {
            "type": "string",
            "title": "Scenario Pack Hash"
          },
          "policy_version": {
            "type": "integer",
            "title": "Policy Version"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioOutcome"
            },
            "type": "array",
            "title": "Scenarios"
          },
          "trace_refs": {
            "items": {
              "$ref": "#/components/schemas/TraceRef"
            },
            "type": "array",
            "title": "Trace Refs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "agent_id",
          "scenario_pack_hash",
          "policy_version",
          "created_at",
          "scenarios",
          "trace_refs"
        ],
        "title": "ReportBody",
        "description": "The bytes that get signed. Fields are the Phase 6 schema."
      },
      "RunAccepted": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "RunAccepted",
        "description": "The id a client polls. The suite has not run yet."
      },
      "RunCreate": {
        "properties": {
          "suite": {
            "type": "string",
            "minLength": 1,
            "title": "Suite"
          },
          "agent": {
            "type": "string",
            "minLength": 1,
            "title": "Agent"
          },
          "policy": {
            "type": "string",
            "minLength": 1,
            "title": "Policy"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "suite",
          "agent",
          "policy"
        ],
        "title": "RunCreate",
        "description": "The suite directory, the agent entry, and the policy file to score with."
      },
      "RunList": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Runs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "runs"
        ],
        "title": "RunList",
        "description": "Runs newest first."
      },
      "RunStatusBody": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status"
          },
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "suite_hash": {
            "type": "string",
            "title": "Suite Hash"
          },
          "policy_hash": {
            "type": "string",
            "title": "Policy Hash"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "status",
          "agent",
          "suite_hash",
          "policy_hash",
          "created_at"
        ],
        "title": "RunStatusBody",
        "description": "What `GET /runs/{id}` returns."
      },
      "RunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status"
          },
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "passed_count": {
            "type": "integer",
            "title": "Passed Count"
          },
          "scenario_count": {
            "type": "integer",
            "title": "Scenario Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "status",
          "agent",
          "created_at",
          "passed_count",
          "scenario_count"
        ],
        "title": "RunSummary",
        "description": "One row on the runs list."
      },
      "RunTraces": {
        "properties": {
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/ScenarioTrace"
            },
            "type": "array",
            "title": "Scenarios"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scenarios"
        ],
        "title": "RunTraces",
        "description": "Traces for a finished run, in scenario id order."
      },
      "ScenarioDiff": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "left_passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Left Passed"
          },
          "right_passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Right Passed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "severity",
          "left_passed",
          "right_passed"
        ],
        "title": "ScenarioDiff",
        "description": "One scenario whose pass/fail is not the same on the two runs."
      },
      "ScenarioOutcome": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "suite": {
            "type": "string",
            "title": "Suite"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "deciding_event": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deciding Event"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "failing_rules": {
            "items": {
              "$ref": "#/components/schemas/FailingRule"
            },
            "type": "array",
            "title": "Failing Rules"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "suite",
          "passed"
        ],
        "title": "ScenarioOutcome",
        "description": "Pass or fail for one scenario, plus the rule ids and excerpt a reviewer needs."
      },
      "ScenarioTrace": {
        "properties": {
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "trace": {
            "$ref": "#/components/schemas/AgentTrace"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scenario_id",
          "trace"
        ],
        "title": "ScenarioTrace",
        "description": "The stored trace for one scenario."
      },
      "SignedReport": {
        "properties": {
          "report": {
            "$ref": "#/components/schemas/ReportBody"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          },
          "algorithm": {
            "type": "string",
            "enum": [
              "local",
              "ML-DSA-44",
              "qknot"
            ],
            "title": "Algorithm"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "report",
          "signature",
          "algorithm"
        ],
        "title": "SignedReport",
        "description": "A report body plus the signature over its canonical bytes."
      },
      "TraceEvent": {
        "properties": {
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "kind": {
            "type": "string",
            "enum": [
              "llm_call",
              "tool_call",
              "tool_result",
              "final"
            ],
            "title": "Kind"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "arguments": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arguments"
          },
          "output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "started_ns": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started Ns"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "token_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Count"
          },
          "args_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Args Sha256"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sequence",
          "kind"
        ],
        "title": "TraceEvent",
        "description": "One recorded moment in a run."
      },
      "TraceRef": {
        "properties": {
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scenario_id",
          "sha256"
        ],
        "title": "TraceRef",
        "description": "Sha256 of one scenario's canonical trace JSON."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}